Update and rename fenwick_tree_2D.cpp to fenwick_tree_2d.cpp
